NonlinearNormalForm

Documentation for NonlinearNormalForm.

NonlinearNormalForm.compose_it!

NonlinearNormalForm.compose_it!

NonlinearNormalForm.gofix

NonlinearNormalForm.compose_it! — Method

compose_it!(m, m2, m1)

Composes the maps in-place, m2 ∘ m1. Aliasing is allowed, however in some cases may be slower than not aliasing, especially when m === m1. Assumes the destination map is properly set up (with correct types promoted if necessary), and that m.x[1:nv] and m.Q.q contain allocated TPSs.

NonlinearNormalForm.compose_it! — Method

compose_it!(m, m2, m1)

Composes the maps in-place, m2 ∘ m1. Aliasing is allowed, however in some cases may be slower than not aliasing, especially when m === m1. Assumes the destination map is properly set up (with correct types promoted if necessary), and that m.x[1:nv] and m.Q.q contain allocated TPSs.

NonlinearNormalForm.gofix — Function

Finds the parameter dependent fixed point to the specified order. Does not currently allow coasting beam.

(a1-I) = (M-I)^-1*(k)

(M-I)(a1-I) = k Ma1 - M - a1 - I = k

M*a1 = k +a1 + M + I

M*a1 = a1+k