NonlinearNormalForm

Documentation for NonlinearNormalForm.

NonlinearNormalForm.S

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.Probe

NonlinearNormalForm.Quaternion

NonlinearNormalForm.SymplecticS

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TaylorMap

NonlinearNormalForm.VectorField

NonlinearNormalForm.VectorField

NonlinearNormalForm.VectorField

Base.:*

Base.:∘

Base.:∘

Base.inv

Base.log

Base.one

Base.one

Base.zero

Base.zero

GTPSA.compose!

GTPSA.compose!

LinearAlgebra.mul!

LinearAlgebra.mul!

NonlinearNormalForm.checksymp

NonlinearNormalForm.compose

NonlinearNormalForm.compose

NonlinearNormalForm.compose_it!

NonlinearNormalForm.compose_it!

NonlinearNormalForm.exp!

NonlinearNormalForm.fgrad

NonlinearNormalForm.inv!

NonlinearNormalForm.is_orbital_resonance

NonlinearNormalForm.is_spin_resonance

NonlinearNormalForm.is_tune_shift

NonlinearNormalForm.lb

NonlinearNormalForm.lb!

NonlinearNormalForm.locate_modes!

NonlinearNormalForm.log!

NonlinearNormalForm.mat_eigen

NonlinearNormalForm.mat_eigen!

NonlinearNormalForm.moveback_unstable!

NonlinearNormalForm.normalize_eigenmode!

NonlinearNormalForm.S — Constant

Generic symplectic skew symmetric S matrix (size inferred from other matrix in operations) using SkewLinearAlgebra's JMatrix

NonlinearNormalForm.DAMap — Type

DAMap{S,T,U,V,W} <: TaylorMap{S,T,U,V,W}

TaylorMap that composes and inverses as a DAMap (with the scalar part ignored). See TaylorMap for more information.

NonlinearNormalForm.DAMap — Method

DAMap(M::AbstractMatrix; use::UseType=GTPSA.desc_current, x0::Vector=zeros(eltype(M), size(M,1)), Q::Union{Quaternion,Nothing}=nothing, E::Union{Matrix,Nothing}=nothing, idpt::Union{Bool,Nothing}=nothing, spin::Union{Bool,Nothing}=nothing, FD::Union{Bool,Nothing}=nothing)

This function could be optimized.

M must represent a matrix with linear indexing.

Constructs a DAMap with the passed matrix of scalars M as the linear part of the TaylorMap, and optionally the entrance coordinates x0, Quaternion for spin Q, and FD matrix E as keyword arguments. The helper keyword arguments spin and FD may be set to true to construct a DAMap with an identity quaternion/FD matrix, or false for no spin/FD. Note that setting spin/FD to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor.

NonlinearNormalForm.DAMap — Method

DAMap(;use::UseType=GTPSA.desc_current, x::Vector=vars(getdesc(use)), x0::Vector=zeros(eltype(eltype(x)), numvars(use)), Q::Union{Quaternion,Nothing}=nothing, E::Union{Matrix,Nothing}=nothing, idpt::Union{Bool,Nothing}=nothing, spin::Union{Bool,Nothing}=nothing, FD::Union{Bool,Nothing}=nothing)

Constructs a DAMap with the passed vector of TPS/ComplexTPS64 as the orbital ray, and optionally the entrance coordinates x0, Quaternion for spin Q, and FD matrix E as keyword arguments. The helper keyword arguments spin and FD may be set to true to construct a DAMap with an identity quaternion/FD matrix, or false for no spin/FD. Note that setting spin/FD to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor. The use kwarg may also be used to change the Descriptor of the TPSs, provided the number of variables

parameters agree (orders may be different).

NonlinearNormalForm.DAMap — Method

DAMap(m::Union{TaylorMap{S,T,U,V,W},Probe{S,T,U,V,W}}; use::UseType=m, idpt::Union{Nothing,Bool}=m.idpt) where {S,T,U,V,W}

Creates a new copy of the passed TaylorMap as a DAMap.

If use is not specified, then the same GTPSA Descriptor as m will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the copy of m as a new DAMap will have the same Descriptor as in use. The total number of variables + parameters must agree, however the orders may be different.

If idpt is not specified, then the same idpt as m will be used, else that specified will be used.

NonlinearNormalForm.Probe — Type

Probe{S,T,U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing},W<:Union{Nothing,Bool}}

Parametric type used for tracking. The orbital/spin part can contain either scalars or TPSs. If spin is included, the field Q is a Quaternion with the same type as x, else Q is nothing. If radiation is included, the field E contains the FD matrix with eltype(E)==S, else E is nothing.

If all planes are exhibiting pseudo-harmonic oscillations, then idpt is nothing. If the last plane is coasting, then idpt specifies which variable in the last plane is constant (energy-like): idpt=false if the variable with index NV-1 is constant, or idpt=true is the variable with index NV is constant.

NonlinearNormalForm.Quaternion — Type

Quaternion{T <: Number}

Lightweight quaternion implementation for simulations. Mutable in order to compose in GTPSA.

NonlinearNormalForm.SymplecticS — Type

Generic symplectic skew symmetric S matrix (size inferred from other matrix in operations) using SkewLinearAlgebra's JMatrix

NonlinearNormalForm.TPSAMap — Type

TPSAMap{S,T,U,V,W} <: TaylorMap{S,T,U,V,W}

TaylorMap that composes and inverses as a TPSAMap (with the scalar part included). See TaylorMap for more information.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(M::AbstractMatrix; use::UseType=GTPSA.desc_current, x0::Vector=zeros(eltype(M), size(M,1)), Q::Union{Quaternion,Nothing}=nothing, E::Union{Matrix,Nothing}=nothing, idpt::Union{Bool,Nothing}=nothing, spin::Union{Bool,Nothing}=nothing, FD::Union{Bool,Nothing}=nothing)

This function could be optimized.

M must represent a matrix with linear indexing.

Constructs a TPSAMap with the passed matrix of scalars M as the linear part of the TaylorMap, and optionally the entrance coordinates x0, Quaternion for spin Q, and FD matrix E as keyword arguments. The helper keyword arguments spin and FD may be set to true to construct a TPSAMap with an identity quaternion/FD matrix, or false for no spin/FD. Note that setting spin/FD to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(;use::UseType=GTPSA.desc_current, x::Vector=vars(getdesc(use)), x0::Vector=zeros(eltype(eltype(x)), numvars(use)), Q::Union{Quaternion,Nothing}=nothing, E::Union{Matrix,Nothing}=nothing, idpt::Union{Bool,Nothing}=nothing, spin::Union{Bool,Nothing}=nothing, FD::Union{Bool,Nothing}=nothing)

Constructs a TPSAMap with the passed vector of TPS/ComplexTPS64 as the orbital ray, and optionally the entrance coordinates x0, Quaternion for spin Q, and FD matrix E as keyword arguments. The helper keyword arguments spin and FD may be set to true to construct a TPSAMap with an identity quaternion/FD matrix, or false for no spin/FD. Note that setting spin/FD to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor. The use kwarg may also be used to change the Descriptor of the TPSs, provided the number of variables

parameters agree (orders may be different).

NonlinearNormalForm.TPSAMap — Method

TPSAMap(m::Union{TaylorMap{S,T,U,V,W},Probe{S,T,U,V,W}}; use::UseType=m, idpt::Union{Nothing,Bool}=m.idpt) where {S,T,U,V,W}

Creates a new copy of the passed TaylorMap as a TPSAMap.

If use is not specified, then the same GTPSA Descriptor as m will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the copy of m as a new TPSAMap will have the same Descriptor as in use. The total number of variables + parameters must agree, however the orders may be different.

If idpt is not specified, then the same idpt as m will be used, else that specified will be used.

NonlinearNormalForm.TaylorMap — Type

TaylorMap{S,T<:TPS,U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing},W<:Union{Nothing,Bool}}

Abstract type for TPSAMap and DAMap used for normal form analysis.

All TaylorMaps contain x0 and x as the entrance coordinates and transfer map as a truncated power series respectively. If spin is included, a field Q containing a Quaternion as a truncated power series is included, else Q is nothing. If stochasticity is included, a field E contains a matrix of the envelope for the FD kicks, else E is nothing.

If all planes are exhibiting pseudo-harmonic oscillations, then idpt is nothing. If the last plane is coasting, then idpt specifies which variable in the last plane is constant (energy-like): idpt=false if the variable with index NV-1 is constant, or idpt=true is the variable with index NV is constant.

Fields

x0 – Entrance coordinates of the map, Taylor expansion point

x – Orbital ray as a truncated power series, expansion around x0 + scalar part equal to EXIT coordinates of map

Q – Quaternion as a truncated power series if spin is included, else nothing

E – Matrix of the envelope for FD kicks if included, else nothing

idpt – If the last plane is coasting, then idpt=false if the first variable in last plane is constant (energy-like) or true if the second. If no coasting, then idpt=nothing

NonlinearNormalForm.VectorField — Type

Lie operator to act on maps. Can be turned into a map with exp(:F:)

NonlinearNormalForm.VectorField — Method

VectorField(m::DAMap)

Creates a VectorField from the map

NonlinearNormalForm.VectorField — Method

VectorField(h::TPS; Q::Union{Quaternion,Nothing}=nothing, spin::Union{Bool,Nothing}=nothing)

Constructs a VectorField from the passed Hamiltonian h. Explicity, for h, constructs a vector field F such that

F.x = [-∂h/∂p₁, ∂h/∂q₁, ...]

Base.:* — Method

*(F::VectorField, m1::DAMap)

Calculates a Lie operator VectorField acting on a DAMap. Explicity, if spin is included that is F * m = (F.x, F.Q) * (m.x, m.Q) = (F.x ⋅ ∇ m.x , F.x ⋅ ∇ m.Q + m.Q*F.Q)

Base.:∘ — Method

∘(m2::DAMap,m1::DAMap) -> DAMap

DAMap composition, ignoring the scalar part of m1

Base.:∘ — Method

∘(m2::TPSAMap,m1::TPSAMap) -> TPSAMap

TPSAMap composition, including the scalar part of m1

Base.inv — Method

inv(m1::TaylorMap; dospin::Bool=true)

Inverts the TaylorMap.

Keyword Arguments

dospin – Specify whether to invert the quaternion as well, default is true

Base.log — Method

log(m1::DAMap)

Computes the log of the map m1 - that is, calculates the VectorField F that would represent the map m1 as a Lie exponent exp(F). The map m1 should be close to the identity for this to converge quickly.

Base.one — Method

one(m::DAMap)

Construct an identity map based on m.

Base.one — Method

one(m::TPSAMap)

Construct an identity map based on m.

Base.zero — Method

zero(m::DAMap)

Creates a DAMap with the same GTPSA Descriptor, and spin/FD on/off, as m but with all zeros for each quantity (except for the immutable parameters in x[nv+1:nn], which will be copied from m.x)

Base.zero — Method

zero(m::TPSAMap)

Creates a TPSAMap with the same GTPSA Descriptor, and spin/FD on/off, as m but with all zeros for each quantity (except for the immutable parameters in x[nv+1:nn], which will be copied from m.x)

GTPSA.compose! — Method

compose!(m::DAMap, m2::DAMap, m1::DAMap; keepscalar::Bool=true, workref::Union{Nothing,Vector{<:Union{Float64,ComplexF64}}}=nothing, dospin::Bool=true, workprom::Union{Nothing,Tuple{Vararg{Vector{<:ComplexTPS64}}}}=prepcompworkprom(m,m2,m1))

In-place DAMap composition which calculates m = m2 ∘ m1, ignoring the scalar part of m1.

Aliasing m === m2 is allowed, but NOT m === m1. m2 === m1 is fine. The destination map m should be properly set up (with correct types promoted if necessary), and have m.x[1:nv] (and m.Q if spin) containing allocated TPSs.

If keep_scalar is set to true, then the scalar part of m1 is retained. In this case, a temporary vector work_ref must be used to store the scalar part before calling the low-level compose_it!, then added back into m1 after composition. work_ref can be optionally provided, or will be created internally in this case. Default is true.

If dospin is true, then the quaternion part of the maps will be composed as well. Default is true.

See the documentation for compose_it! for information on work_prom.

Keyword Arguments

keep_scalar – Specify whether to keep the scalar part in m1 or throw it away. If true, a temporary vector storing the scalar part must be used. Default is true

work_ref – If keep_scalar is true, the temporary vector can be provided in this keyword argument. If nothing is provided, the temporary will be created internally. Default is nothing

dospin – Specify whether or not to include the quaternions in the concatenation. Default is true

work_prom – Temporary vector of allocated ComplexTPS64s when there is implicit promotion. See the compose_it! documentation for more details. Default is output from prep_comp_work_prom(m, m2, m1)

GTPSA.compose! — Method

compose!(m::TPSAMap, m2::TPSAMap, m1::TPSAMap; dospin::Bool=true, work_prom::Union{Nothing,Tuple{Vararg{Vector{<:ComplexTPS64}}}}=prep_comp_work_prom(m,m2,m1))

In-place TPSAMap composition which calculates m = m2 ∘ m1, including the scalar part of m1.

Aliasing m === m2 is allowed, but NOT m === m1. m2 === m1 is fine. The destination map m should be properly set up (with correct types promoted if necessary), and have m.x[1:nv] (and m.Q if spin) containing allocated TPSs.

If dospin is true, then the quaternion part of the maps will be composed as well. Default is true.

See the documentation for compose_it! for information on work_prom.

Keyword Arguments

dospin – Specify whether or not to include the quaternions in the concatenation. Default is true

work_prom – Temporary vector of allocated ComplexTPS64s when there is implicit promotion. See the compose_it! documentation for more details. Default is output from prep_comp_work_prom(m, m2, m1)

LinearAlgebra.mul! — Method

mul!(m::DAMap, F::VectorField, m1::DAMap; work_Q::Union{Quaternion,Nothing}=prep_work_Q(F))

Computes the Lie operator F acting on a DAMap m1, and stores the result in m. Explicity, that is F * m = (F.x, F.Q) * (m.x, m.Q) = (F.x ⋅ ∇ m.x , F.x ⋅ ∇ m.Q + m.Q*F.Q)

Keyword Arguments- work_Q – Quaternion{T} if spin is included in the vector field, else nothing

LinearAlgebra.mul! — Method

mul!(Q::Quaternion, Q1::Quaternion, Q2::Quaternion)

Sets Q equal to Q1*Q2, with zero allocations. Q must not be aliased with either Q1 or Q2.

NonlinearNormalForm.checksymp — Method

checksymp(M)

Returns tranpose(M)*S*M - S, where S is the skew-symmetric matrix S = blkdiag([0 1; -1 0], ...). If M is symplectic, then the result should be a matrix containing all zeros. The non-symplectic parts of the matrix can be identified by those nonzero elements in the result.

NonlinearNormalForm.compose — Method

compose(m2::DAMap,m1::DAMap) -> DAMap

DAMap composition, which calculates m2 ∘ m1 ignoring the scalar part of m1

NonlinearNormalForm.compose — Method

compose(m2::TPSAMap,m1::TPSAMap) -> TPSAMap

TPSAMap composition, which calculates m2 ∘ m1 including the scalar part of m1

NonlinearNormalForm.compose_it! — Method

compose_it!(m, m2, m1; dospin::Bool=true, dostochastic::Bool=true, work_prom::Union{Nothing,Tuple{Vararg{Vector{<:ComplexTPS64}}}}=prep_comp_work_prom(m,m2,m1))

Low level composition function, m = m2 ∘ m1. Aliasing m with m2 is allowed, but not m with m1. Assumes the destination map is properly set up (with correct types promoted if necessary), and that m.x[1:nv] (and m.Q if spin) contain allocated TPSs. The parameters part of m.x (m.x[nv+1:nn]) does not need to contain allocated TPSs.

If promotion is occuring, then one of the maps is ComplexTPS64 and the other TPS, with output map ComplexTPS64. Note that the spin part is required to agree with the orbital part in terms of type by definition of the TaylorMap struct. work_prom can optionally be passed as a tuple containing the temporary ComplexTPS64s if promotion is occuring:

If eltype(m.x) != eltype(m1.x) (then m1 must be promoted): work_prom[1] = m1x_prom # Length >= nv+np, Vector{ComplexTPS64}

else if eltype(m.x) != eltype(m2.x) (then m2 must be promoted):

work_prom[1] = m2x_prom  # Length >= nv, Vector{ComplexTPS64}
work_prom[2] = m2Q_prom  # Length >= 4, Vector{ComplexTPS64}

Note that the ComplexTPS64s in the vector(s) must be allocated and have the same Descriptor.

If spin is included, not that the final quaternion concatenation step mul! will create allocations

NonlinearNormalForm.compose_it! — Method

compose_it!(m, m2, m1; dospin::Bool=true, dostochastic::Bool=true, work_prom::Union{Nothing,Tuple{Vararg{Vector{<:ComplexTPS64}}}}=prep_comp_work_prom(m,m2,m1))

Low level composition function, m = m2 ∘ m1. Aliasing m with m2 is allowed, but not m with m1. Assumes the destination map is properly set up (with correct types promoted if necessary), and that m.x[1:nv] (and m.Q if spin) contain allocated TPSs. The parameters part of m.x (m.x[nv+1:nn]) does not need to contain allocated TPSs.

If promotion is occuring, then one of the maps is ComplexTPS64 and the other TPS, with output map ComplexTPS64. Note that the spin part is required to agree with the orbital part in terms of type by definition of the TaylorMap struct. work_prom can optionally be passed as a tuple containing the temporary ComplexTPS64s if promotion is occuring:

If eltype(m.x) != eltype(m1.x) (then m1 must be promoted): work_prom[1] = m1x_prom # Length >= nv+np, Vector{ComplexTPS64}

else if eltype(m.x) != eltype(m2.x) (then m2 must be promoted):

work_prom[1] = m2x_prom  # Length >= nv, Vector{ComplexTPS64}
work_prom[2] = m2Q_prom  # Length >= 4, Vector{ComplexTPS64}

Note that the ComplexTPS64s in the vector(s) must be allocated and have the same Descriptor.

If spin is included, not that the final quaternion concatenation step mul! will create allocations

NonlinearNormalForm.exp! — Method

exp!(m::DAMap, F::VectorField, m1::DAMap; work_maps::Tuple{Vararg{DAMap}}=(zero(m1),zero(m1)), work_Q::Union{Quaternion,Nothing}=prep_work_Q(F))

Computes exp(F)*m1, and stores the result in m. Explicity, this is exp(F)*m1 = m1 + F*m1 + 1/2*F*(F*m1) + 1/6*F*(F*(F*m1)) + ..., where * is the operation of a VectorField on the map. See the documentation for mul! for more details of this operation. If the linear part of F is zero, then the number of iterations is equal to the number of higher orders left.

Keyword Arguments

work_maps – Tuple of 2 DAMaps of the same type as m1

work_Q – Quaternion{T} if spin is included in the vector field, else nothing

NonlinearNormalForm.fgrad — Method

fgrad(F::AbstractVector{<:T}, h::T) where {T<:TPS{<:Union{Float64,ComplexF64}}}

Calculates F⋅∇h.

NonlinearNormalForm.inv! — Method

inv!(m::TaylorMap, m1::TaylorMap; dospin::Bool=true, work_ref::Union{Nothing,Vector{<:Union{Float64,ComplexF64}}}=nothing)

In-place inversion of the TaylorMap setting m = inv(m1). Aliasing m === m1 is allowed, however in this case a temporary vector must be used to store the scalar part of m1 prior to inversion so that the entrance/exit coordinates of the map can be properly handled.

Keyword Arguments

dospin – Specify whether to invert the quaternion as well, default is true

work_ref – If m === m1, then a temporary vector must be used to store the scalar part. If not provided and m === m1, this temporary will be created internally. Default is nothing

NonlinearNormalForm.is_orbital_resonance — Method

is_orbital_resonance(varidx, ords, nhv, res)

Checks if the monomial corresponds to a particular resonance (and resonance in the same family)

Note that for 3Q_x = integer, we also have 6Q_x though where

lambda(3Q_x) = integer where lambda*3 <= MAXORD + 1 according to notes are in the same family and so these must be left in too. These are the same family

each column of resonance corresponds to one res in the family

Each row is a multiple of previous

So For example for the 1Q_1 + 2Q2 + 3*Q3 = integer resonance: m = [ 1 2 3 ...; 2 4 6 ...; 3 6 9 ...];

NonlinearNormalForm.is_spin_resonance — Method

nu_s + j dot Q = n

0Qx + 1Qy + 2*Qs = n

IMPORTANT:::: For spin resonances, there is only one resonance in each resonance family:

m = [0 ; 1 ]

ms = [1];

In the code, check - of m + ms

NonlinearNormalForm.is_tune_shift — Function

is_tune_shift(varidx, ords, nhv)

Checks if the monomial corresponds to a tune shift.

Input

varidx – Current variable index (e.g. 1 is x, 2 is px, etc)

ords – Array of monomial index as orders

nhv – Number harmonic variables

hamiltonian – Default is false, if the monomial is in a vector field and not a hamitlonian then this should be false.

NonlinearNormalForm.lb! — Method

lb!(G::VectorField, F::VectorField, H::VectorField; work_Q::Union{Quaternion,Nothing}=nothing)

Computes the Lie bracket of the vector fields F and H, and stores the result in G. See lb for more details.

Keyword Arguments

work_Q – Quaternion{T} if spin is included in the vector field, else nothing

NonlinearNormalForm.lb — Method

lb(F::VectorField, H::VectorField)

Computes the Lie bracket of the vector fields F and H. Explicitly, and including spin (with the lower case letter for the quaternion of the vector field), this is:

(G,g) = ⟨(F,f) , (H,h)⟩ = (F⋅∇H-H⋅∇F , [h,f]+F⋅∇h-G⋅∇f)

where [h,f] = h*f - f*h is just a quaternion commutator. See Equation 44.52 in the Bmad manual

NonlinearNormalForm.locate_modes! — Function

locate_modes!(evecs, evals=nothing; sort=true, modes=nothing)

For each mode (every pair of rows in evecs), determines the eigenvector (each column in evecs) with the largest norm in that mode. The eigenvectors must have dimensionality of 2n, n ∈ ℤ and be next to each other pairs.

If sort is true, then the pairs of eigenvectors are sorted in-place. The eigenvalues will also be sorted if evals is provided. If sort is false, then the modes vector is filled with a mapping of the current eigenvector position to its mode (index of modes is the current eigenvector position, and the element at that index in modes is the mode that it corresponds to). For sort=true, modes == 1:n after calling this function.

If the mode locating is successful, true is returned. For unsuccessful mode locating, false is returned, modes will contain junk, and the eigenvectors/values may be partially sorted.

Assumes the eigenvectors (and eigenvalues if provided) are already in pairs (next to each other) starting at the first column.

Input:

evecs – 2n x 2n matrix of eigenvectors already in pairs

evals – Vector of length 2n containing the eigenvalues corresponding to the eigenvectors in evecs

sort – (Optional) kwarg to specify whether or not to sort the eigenvectors, default is true

Output:

ret – Returns true if the mode locating is successful, false if otherwise

modes – (Optional) kwarg for eigvec -> mode mapping, length(evecs) == Int(size(evecs,1)/2)

NonlinearNormalForm.log! — Method

log!(F::VectorField, m1::DAMap; work::Tuple{DAMap,DAMap,DAMap,VectorField,VectorField}=prep_log_work(m1), work_Q::Union{Quaternion,Nothing}=prep_work_Q(F))

Computes the log of the map m1 - that is, calculates the VectorField F that would represent the map m1 as a Lie exponent exp(F) - and stores the result in F. The map m1 should be close to the identity for this to converge quickly.

Keyword Arguments

work – Tuple of 3 DAMaps of the same type as m1 followed by 2 VectorFieldS

work_Q – Quaternion{T} if spin is included in the vector field, else nothing

NonlinearNormalForm.mat_eigen! — Method

mat_eigen!(mat; sort=true, phase_modes=true)

Same as mat_eigen, but mutates mat for speed. See the documentation for mat_eigen for more details.

NonlinearNormalForm.mat_eigen — Method

mat_eigen(mat; sort=true, phase_modes=true)

Given a matrix mat with an even number of rows/cols, calculates the eigenvectors and eigenvalues. For stable eigenmodes (complex conjugate eigenvector/eigenvalue pairs), the eigenvectors are normalized so that vⱼ'*S*vⱼ = +im for odd j, and -im for even j.

If sort is true, then each eigenvector/eigenvalue pair will be sorted according to the mode it best identifies with. A warning will be printed if the mode sorting fails. Mode sorting will automatically fail if more than 1 mode is unstable. Default is true.

If phase_modes is true, then each stable mode will be multiplied by a phase factor to make the first component of that mode (e.g. the x, y, or z component, NOT the px, py, or pz component) be fully real. This both makes the eigenvector "pretty", and in this case of weak coupling, ensures that vⱼ'*vⱼ₊₁ for j=(1, 3, 5) has a positive imaginary part so that the eigenvectors/eigenvalues for odd j are associated with the positive tune and even j the negative tune. This phase factor is harmless/useless to include in a highly-coupled matrix.

For complex matrices, Julia's eigen, which is called by mat_eigen, is type-unstable.

NonlinearNormalForm.moveback_unstable! — Method

moveback_unstable!(F::Eigen) -> Int

Moves back eigenvectors with eigenvalues having a zero imaginary component to the end of the values and vectors arrays in the Eigen struct, and returns the number of unstable eigenvectors.

Note that if more than 1 mode is unstable, the pair of eigenvectors corresponding to a mode will not necessarily be next to each other at the end of the matrix.

NonlinearNormalForm.normalize_eigenmode! — Function

normalize_eigenmode!(evec_pair, eval_pair, phase_mode::Integer=-1)

Normalizes the complex-conjugate eigenvector pair so that vⱼ'*S*vⱼ = +im for j=1, and -im for j=2. This may involve swapping the complex-conjugate eigenvectors/eigenvalues.

If phase_mode is set to a particular mode (1, 2, 3), then a phase will be multiplied by the eigenvectors to make them "pretty"; the phase will make the first component of that mode (e.g. the x, y, or z component) be fully real. This gives the eigenvectors a simple form in the weakly-coupled case, with the product v₁'*v₂ having a positive imaginary part, and ensures the odd numbered eigenvector/eigenvalues are associated with the tune and the even numbered ones have the negative of the tune. In cases with a lot of coupling, there is no simple way to define the positive or negative tune, and multiplying by this phase factor is harmless. See the "Tunes From One-Turn Matrix Eigen Analysis" section in the Bmad manual for more details.