NonlinearNormalForm

Documentation for NonlinearNormalForm.

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TaylorMap

Base.inv

NonlinearNormalForm.checksymp

NonlinearNormalForm.compose_it!

NonlinearNormalForm.compose_it!

NonlinearNormalForm.exppb

NonlinearNormalForm.fgrad

NonlinearNormalForm.gethamiltonian

NonlinearNormalForm.getvectorfield

NonlinearNormalForm.lb

NonlinearNormalForm.locate_planes!

NonlinearNormalForm.logpb

NonlinearNormalForm.mat_eigen

NonlinearNormalForm.mat_eigen!

NonlinearNormalForm.moveback_unstable!

NonlinearNormalForm.normalize_evecs!

NonlinearNormalForm.pb

NonlinearNormalForm.ptinv

NonlinearNormalForm.DAMap — Type

DAMap{S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix{S},Nothing}} <: TaylorMap{S,T,U,V}

TaylorMap that composes and inverses as a DAMap (with the scalar part ignored). See TaylorMap for more information.

NonlinearNormalForm.DAMap — Method

DAMap(M::Matrix{S}; x0::Vector{S}=zeros(numtype(first(x)), numvars(x)), Q::U=nothing, E::V=nothing,  spin::Union{Bool,Nothing}=nothing, radiation::Union{Bool,Nothing}=nothing, use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Constructs a DAMap with the passed matrix of scalars as the linear part of the TaylorMap, and optionally the entrance coordinates x0, Quaternion for spin Q, and stochastic matrix E as keyword arguments. The helper keyword arguments spin and radiation may be set to true to construct a DAMap with an identity quaternion/stochastic matrix, or false for no spin/radiation. Note that setting spin/radiation to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor.

NonlinearNormalForm.DAMap — Method

DAMap(p::Probe{S,T,U,V}; use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Creates a DAMap from the Probe, which must contain TPSs.

If use is not specified, then the same GTPSA Descriptor as p will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the p promoted to a DAMap will have the same Descriptor as in use. The number of variables and parameters must agree, however the orders may be different.

NonlinearNormalForm.DAMap — Method

DAMapTaylorMap{S,T,U,V}; use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Creates a new copy of the passed TaylorMap as a DAMap.

If use is not specified, then the same GTPSA Descriptor as m will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the copy of m as a new DAMap will have the same Descriptor as in use. The number of variables and parameters must agree, however the orders may be different.

NonlinearNormalForm.DAMap — Method

DAMap(u::UndefInitializer, m::TaylorMap{S,T,U,V}) where {S,T,U,V}

Creates an undefined DAMap based on m (same Descriptor and with radiation/spin on/off.)

NonlinearNormalForm.DAMap — Method

DAMap(x::Vector{T}=zeros(TPS, numvars(GTPSA.desc_current)); x0::Vector{S}=zeros(numtype(first(x)), numvars(x)), Q::U=nothing, E::V=nothing,  spin::Union{Bool,Nothing}=nothing, radiation::Union{Bool,Nothing}=nothing, use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Constructs a DAMap with the passed vector of TPS/ComplexTPS as the orbital ray, and optionally the entrance coordinates x0, Quaternion for spin Q, and stochastic matrix E as keyword arguments. The helper keyword arguments spin and radiation may be set to true to construct a DAMap with an identity quaternion/stochastic matrix, or false for no spin/radiation. Note that setting spin/radiation to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor. The use kwarg may also be used to change the Descriptor of the TPSs, provided the number of variables and parameters agree (orders may be different).

NonlinearNormalForm.TPSAMap — Type

TPSAMap{S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix{S},Nothing}} <: TaylorMap{S,T,U,V}

TaylorMap that composes and inverses as a TPSAMap (with the scalar part included). See TaylorMap for more information.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(M::Matrix{S}; x0::Vector{S}=zeros(numtype(first(x)), numvars(x)), Q::U=nothing, E::V=nothing,  spin::Union{Bool,Nothing}=nothing, radiation::Union{Bool,Nothing}=nothing, use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Constructs a TPSAMap with the passed matrix of scalars as the linear part of the TaylorMap, and optionally the entrance coordinates x0, Quaternion for spin Q, and stochastic matrix E as keyword arguments. The helper keyword arguments spin and radiation may be set to true to construct a TPSAMap with an identity quaternion/stochastic matrix, or false for no spin/radiation. Note that setting spin/radiation to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(p::Probe{S,T,U,V}; use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Creates a TPSAMap from the Probe, which must contain TPSs.

If use is not specified, then the same GTPSA Descriptor as p will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the p promoted to a TPSAMap will have the same Descriptor as in use. The number of variables and parameters must agree, however the orders may be different.

NonlinearNormalForm.TPSAMap — Method

TPSAMapTaylorMap{S,T,U,V}; use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Creates a new copy of the passed TaylorMap as a TPSAMap.

If use is not specified, then the same GTPSA Descriptor as m will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the copy of m as a new TPSAMap will have the same Descriptor as in use. The number of variables and parameters must agree, however the orders may be different.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(u::UndefInitializer, m::TaylorMap{S,T,U,V}) where {S,T,U,V}

Creates an undefined TPSAMap based on m (same Descriptor and with radiation/spin on/off.)

NonlinearNormalForm.TPSAMap — Method

TPSAMap(x::Vector{T}=zeros(TPS, numvars(GTPSA.desc_current)); x0::Vector{S}=zeros(numtype(first(x)), numvars(x)), Q::U=nothing, E::V=nothing,  spin::Union{Bool,Nothing}=nothing, radiation::Union{Bool,Nothing}=nothing, use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Constructs a TPSAMap with the passed vector of TPS/ComplexTPS as the orbital ray, and optionally the entrance coordinates x0, Quaternion for spin Q, and stochastic matrix E as keyword arguments. The helper keyword arguments spin and radiation may be set to true to construct a TPSAMap with an identity quaternion/stochastic matrix, or false for no spin/radiation. Note that setting spin/radiation to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor. The use kwarg may also be used to change the Descriptor of the TPSs, provided the number of variables and parameters agree (orders may be different).

NonlinearNormalForm.TaylorMap — Type

TaylorMap{S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix{S},Nothing}}

Abstract type for TPSAMap and DAMap used for normal form analysis.

All TaylorMaps contain x0 and x as the entrance coordinates and transfer map as a truncated power series respectively. If spin is included, a field Q containing a Quaternion as a truncated power series is included, else Q is nothing. If radiation is included, a field E contains a matrix of the envelope for stochastic radiation, else E is nothing.

Fields

x0 – Entrance coordinates of the map, Taylor expansion point

x – Orbital ray as a truncated power series, expansion around x0 + scalar part equal to EXIT coordinates of map

Q – Quaternion as a truncated power series if spin is included, else nothing

E – Matrix of the envelope for stochastic radiation if included, else nothing

Base.inv — Method

inv(ma::Vector{<:Union{TPS,ComplexTPS}})

Inverts the map ma such that ma ∘ inv(ma) = 1 in the variables.

Example

julia> d = Descriptor(2,10); x = vars()[1]; p = vars()[2];

julia> time = 0.01; k = 2; m = 0.01;

julia> h = p^2/(2m) + 1/2*k*x^2;

julia> hf = getvectorfield(h);

julia> map = exppb(-time*hf, [x, p]);

julia> map ∘ inv(map)
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:   1.0000000000000000e+00      1      1   0
-------------------------------------------------
   2:   1.0000000000000002e+00      1      0   1

NonlinearNormalForm.checksymp — Method

checksymp(M::Matrix{T}) where T<:Number

Returns tranpose(M)*S*M - S, where S is the skew-symmetric matrix S = [0 I; -I 0]. If M is symplectic, then the result should be a matrix containing all zeros. The non-symplectic parts of the matrix can be identified by those nonzero elements in the result.

NonlinearNormalForm.compose_it! — Method

compose_it!(m, m2, m1; dospin=true, work_low=nothing, work_prom=nothing)

Low level composition function, m = m2 ∘ m1. Aliasing m with m2 is allowed, but not m with m1. Assumes the destination map is properly set up (with correct types promoted if necessary), and that m.x[1:nv] (and m.Q.q if spin) contain allocated TPSs. The parameters part of m.x (m.x[nv+1:nn]) does not need to contain allocated TPSs.

For all compositions, 5 temporary vectors must be generated that contain Ptr{RTPSA} or Ptr{CTPSA} for each TPS in the map (depending on output type), to pass to the low-level C composition function in GTPSA. Three vectors are for the orbital part (m.x, m2.x, m1.x correspondingly referred in the code as outx_low, m2x_low, and m1x_low) and two are for the spin part (m.Q.q and m2.Q.q correspondingly referred in the code as outQ_low and m2Q_low). These 5 temporaries can be optionally passed as a tuple in work_low, and must satisfy the following requirements:

work_low[1] = outx_low   # Length >= nv
work_low[2] = m2x_low    # Length >= nv
work_low[3] = m1x_low    # Length >= nv+np
work_low[4] = outQ_low   # Length >= 4, could be = work_low[1] or work_low[2] if nv >= 4
work_low[5] = m2Q_low    # Length >= 4, could be = work_low[1] or work_low[2] if nv >= 4

Furthermore, for the spin part both of outx_low and m2x_low could be reused for outQ_low and m2Q_low if nv >= 4 , however m1x_low MUST NOT BE REUSED!

If promotion is occuring, then one of the maps is ComplexTPS and the other TPS, with output map ComplexTPS. Note that the spin part is required to agree with the orbital part in terms of type by definition of the TaylorMap struct. work_prom can optionally be passed as a tuple containing the temporary ComplexTPSs if promotion is occuring:

If eltype(m.x) != eltype(m1.x) (then m1 must be promoted): work_prom[1] = m1x_prom # Length >= nv+np, Vector{ComplexTPS}

else if eltype(m.x) != eltype(m2.x) (then m2 must be promoted):

work_prom[1] = m2x_prom  # Length >= nv, Vector{ComplexTPS}
work_prom[2] = m2Q_prom  # Length >= 4, Vector{ComplexTPS}

Note that the ComplexTPSs in the vector(s) must be allocated and have the same Descriptor.

NonlinearNormalForm.compose_it! — Method

compose_it!(m, m2, m1; dospin=true, work_low=nothing, work_prom=nothing)

Low level composition function, m = m2 ∘ m1. Aliasing m with m2 is allowed, but not m with m1. Assumes the destination map is properly set up (with correct types promoted if necessary), and that m.x[1:nv] (and m.Q.q if spin) contain allocated TPSs. The parameters part of m.x (m.x[nv+1:nn]) does not need to contain allocated TPSs.

For all compositions, 5 temporary vectors must be generated that contain Ptr{RTPSA} or Ptr{CTPSA} for each TPS in the map (depending on output type), to pass to the low-level C composition function in GTPSA. Three vectors are for the orbital part (m.x, m2.x, m1.x correspondingly referred in the code as outx_low, m2x_low, and m1x_low) and two are for the spin part (m.Q.q and m2.Q.q correspondingly referred in the code as outQ_low and m2Q_low). These 5 temporaries can be optionally passed as a tuple in work_low, and must satisfy the following requirements:

work_low[1] = outx_low   # Length >= nv
work_low[2] = m2x_low    # Length >= nv
work_low[3] = m1x_low    # Length >= nv+np
work_low[4] = outQ_low   # Length >= 4, could be = work_low[1] or work_low[2] if nv >= 4
work_low[5] = m2Q_low    # Length >= 4, could be = work_low[1] or work_low[2] if nv >= 4

Furthermore, for the spin part both of outx_low and m2x_low could be reused for outQ_low and m2Q_low if nv >= 4 , however m1x_low MUST NOT BE REUSED!

If promotion is occuring, then one of the maps is ComplexTPS and the other TPS, with output map ComplexTPS. Note that the spin part is required to agree with the orbital part in terms of type by definition of the TaylorMap struct. work_prom can optionally be passed as a tuple containing the temporary ComplexTPSs if promotion is occuring:

If eltype(m.x) != eltype(m1.x) (then m1 must be promoted): work_prom[1] = m1x_prom # Length >= nv+np, Vector{ComplexTPS}

else if eltype(m.x) != eltype(m2.x) (then m2 must be promoted):

work_prom[1] = m2x_prom  # Length >= nv, Vector{ComplexTPS}
work_prom[2] = m2Q_prom  # Length >= 4, Vector{ComplexTPS}

Note that the ComplexTPSs in the vector(s) must be allocated and have the same Descriptor.

NonlinearNormalForm.exppb — Function

exppb(F::Vector{<:Union{TPS,ComplexTPS}}, m::Vector{<:Union{TPS,ComplexTPS}}=vars(first(F)))

Calculates exp(F⋅∇)m = m + F⋅∇m + (F⋅∇)²m/2! + .... If m is not provided, it is assumed to be the identity.

Example

julia> d = Descriptor(2,10); x = vars()[1]; p = vars()[2];

julia> time = 0.01; k = 2; m = 0.01;

julia> h = p^2/(2m) + 1/2*k*x^2;

julia> hf = getvectorfield(h);

julia> map = exppb(-time*hf, [x, p])
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:   9.9001665555952290e-01      1      1   0
   1:   9.9666999841313930e-01      1      0   1
-------------------------------------------------
   2:  -1.9933399968262787e-02      1      1   0
   2:   9.9001665555952378e-01      1      0   1

NonlinearNormalForm.fgrad — Method

fgrad(F::Vector{<:Union{TPS,ComplexTPS}}, g::Union{TPS,ComplexTPS})

Calculates F⋅∇g.

NonlinearNormalForm.gethamiltonian — Method

gethamiltonian(F::Vector{<:Union{TPS,ComplexTPS}})

Assuming the variables in the TPSA are canonically-conjugate, and ordered so that the canonically- conjugate variables are consecutive (q₁, p₁, q₂, p₂, ...), this function calculates the Hamiltonian from a vector field F that can be obtained from a Hamiltonian (e.g. by getvectorfield). Explicitly, ∫ F₁ dp₁ - ∫ F₂ dq₁ + ... + ∫ F₂ₙ₋₁ dpₙ - ∫ F₂ₙ dqₙ

Example

julia> d = Descriptor(2,10); x = vars();

julia> h = (x[1]^2 + x[2]^2)/2
TPS:
 Coefficient                Order   Exponent
  5.0000000000000000e-01      2      2   0
  5.0000000000000000e-01      2      0   2

julia> F = getvectorfield(h)
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:  -1.0000000000000000e+00      1      0   1
-------------------------------------------------
   2:   1.0000000000000000e+00      1      1   0

julia> gethamiltonian(F)
TPS:
 Coefficient                Order   Exponent
  5.0000000000000000e-01      2      2   0
  5.0000000000000000e-01      2      0   2

NonlinearNormalForm.getvectorfield — Method

getvectorfield(h::Union{TPS,ComplexTPS})::Vector{<:typeof(h)}

Assuming the variables in the TPSA are canonically-conjugate, and ordered so that the canonically- conjugate variables are consecutive (q₁, p₁, q₂, p₂, ...), calculates the vector field (Hamilton's equations) from the passed Hamiltonian, defined as [∂h/∂p₁, -∂h/∂q₁, ...]

Example

julia> d = Descriptor(2,10); x = vars();

julia> h = (x[1]^2 + x[2]^2)/2
TPS:
 Coefficient                Order   Exponent
  5.0000000000000000e-01      2      2   0
  5.0000000000000000e-01      2      0   2

julia> getvectorfield(h)
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:  -1.0000000000000000e+00      1      0   1
-------------------------------------------------
   2:   1.0000000000000000e+00      1      1   0

NonlinearNormalForm.lb — Method

lb(A::Vector{<:Union{TPS,ComplexTPS}}, F::Vector{<:Union{TPS,ComplexTPS}})

Computes the Lie bracket of the vector functions A and F, defined over N variables as Σᵢᴺ Aᵢ (∂F/∂xᵢ) - Fᵢ (∂A/∂xᵢ)

Example

julia> d = Descriptor(2,10); x = vars();

julia> A = [-x[2], x[1]]
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:  -1.0000000000000000e+00      1      0   1
-------------------------------------------------
   2:   1.0000000000000000e+00      1      1   0

julia> F = [-x[1]^2, 2*x[1]*x[2]]
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:  -1.0000000000000000e+00      2      2   0
-------------------------------------------------
   2:   2.0000000000000000e+00      2      1   1

julia> lb(A,F)
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:   4.0000000000000000e+00      2      1   1
-------------------------------------------------
   2:   3.0000000000000000e+00      2      2   0
   2:  -2.0000000000000000e+00      2      0   2

NonlinearNormalForm.locate_planes! — Function

locate_planes!(evecs, evals=nothing; sort=true, planes=nothing)

For each plane (every pair of rows in evecs), determines the eigenvector (each column in evecs) with the largest norm in that plane. The eigenvectors must have dimensionality of 2n, n ∈ ℤ and be next to each other in complex-conjugate pairs.

If sort is true, then the pairs of eigenvectors are sorted in-place. The eigenvalues will also be sorted if evals is provided. If sort is false, then the planes vector is filled with a mapping of the current eigenvector position to its plane (index of planes is the current eigenvector position, and the element at that index in planes is the plane that it corresponds to). For sort=true, planes == 1:n after calling this function.

If the plane locating is successful, true is returned. For unsuccessful plane locating, false is returned, planes will contain junk, and the eigenvectors/values may be partially sorted.

Assumes the eigenvectors (and eigenvalues if provided) are already in pairs (next to each other) starting at the first column.

Input:

evecs – 2n x 2n matrix of eigenvectors already in pairs

evals – Vector of length 2n containing the eigenvalues corresponding to the eigenvectors in evecs

sort – (Optional) kwarg to specify whether or not to sort the eigenvectors, default is true

Output:

ret – Returns true if the plane locating is successful, false if otherwise

planes – (Optional) kwarg for eigvec -> plane mapping, length(evecs) == Int(size(evecs,1)/2)

NonlinearNormalForm.logpb — Function

logpb(mf::Vector{<:Union{TPS,ComplexTPS}}, mi::Vector{<:Union{TPS,ComplexTPS}}=vars(first(F)))

Given a final map mf and initial map mi, this function calculates the vector field F such that mf=exp(F⋅∇)mi. If mi is not provided, it is assumed to be the identity.

julia> d = Descriptor(2,10); x = vars()[1]; p = vars()[2];

julia> time = 0.01; k = 2; m = 0.01;

julia> h = p^2/(2m) + 1/2*k*x^2;

julia> hf = getvectorfield(h);

julia> map = exppb(-time*hf);

julia> logpb(map) == -time*hf
true

NonlinearNormalForm.mat_eigen! — Method

mat_eigen!(mat; sort=true)

Same as mat_eigen, but mutates mat for speed.

NonlinearNormalForm.mat_eigen — Method

mat_eigen(mat; sort=true)

Given a square matrix mat with an even number of rows/cols, calculates the eigenvectors and eigenvalues. For complex conjugate pairs, the eigenvectors are normalized so that conj(vⱼ)*S*vⱼ = +im for odd j, and -im for even j. Unstable eigenvectors/values (with eigenvalues having a zero imaginary component) are moved to the end and normalized to unity.

NonlinearNormalForm.moveback_unstable! — Method

moveback_unstable!(F::Eigen) -> Int

This function moves back eigenvectors with eigenvalues having a zero imaginary component to the end of the values and vectors arrays in the Eigen struct, and returns the number of unstable eigenvectors.

NonlinearNormalForm.normalize_evecs! — Method

normalize_evecs!(evecs)

Assuming the eigenvectors are in complex-conjugate pais, the eigenvectors are normalized so that conj(vⱼ)*S*vⱼ = +im for odd j, and -im for even j where S is the skew-symmetric matrix e.g. [0 1; -1 0] for 2D.

NonlinearNormalForm.pb — Method

pb(f::Union{TPS, ComplexTPS}, g::Union{TPS, ComplexTPS})

Assuming the variables in the TPSA are canonically-conjugate, and ordered so that the canonically- conjugate variables are consecutive (q₁, p₁, q₂, p₂, ...), computes the Poisson bracket of the scalar functions f and g. The Poisson bracket of two functions {f, g} is defined as Σᵢ (∂f/∂qᵢ)(∂g/∂pᵢ) - (∂g/∂qᵢ)(∂f/∂pᵢ).

Examples

julia> d = Descriptor(4,10);

julia> x = vars(d);

julia> f = (x[1]^2 + x[2]^2)/2 + (x[3]^2 + x[4]^2)/2;

julia> pb(f,x[1])
TPS:
  Coefficient              Order     Exponent
  -1.0000000000000000e+00    1        0    1    0    0

julia> pb(f,x[2])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        1    0    0    0

julia> pb(f,x[3])
TPS:
  Coefficient              Order     Exponent
  -1.0000000000000000e+00    1        0    0    0    1

julia> pb(f,x[4])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        0    0    1    0

NonlinearNormalForm.ptinv — Method

ptinv(ma::Vector{<:Union{TPS,ComplexTPS}}, vars::Vector{<:Integer})

Partially-inverts the map ma, inverting only the variables specified by index in vars.