NonlinearNormalForm

Documentation for NonlinearNormalForm.

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.DAMap

NonlinearNormalForm.Quaternion

NonlinearNormalForm.S

NonlinearNormalForm.SymplecticS

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TPSAMap

NonlinearNormalForm.TaylorMap

NonlinearNormalForm.VectorField

NonlinearNormalForm.VectorField

Base.:∘

Base.:∘

Base.inv

Base.log

Base.one

Base.one

Base.zero

Base.zero

GTPSA.compose!

GTPSA.compose!

LinearAlgebra.mul!

LinearAlgebra.mul!

NonlinearNormalForm.checksymp

NonlinearNormalForm.compose

NonlinearNormalForm.compose

NonlinearNormalForm.compose_it!

NonlinearNormalForm.compose_it!

NonlinearNormalForm.exppb

NonlinearNormalForm.fgrad

NonlinearNormalForm.gethamiltonian

NonlinearNormalForm.inv!

NonlinearNormalForm.lb

NonlinearNormalForm.lb!

NonlinearNormalForm.locate_modes!

NonlinearNormalForm.log!

NonlinearNormalForm.log!

NonlinearNormalForm.mat_eigen

NonlinearNormalForm.mat_eigen!

NonlinearNormalForm.moveback_unstable!

NonlinearNormalForm.normalize_eigenmode!

NonlinearNormalForm.pb

NonlinearNormalForm.pb!

NonlinearNormalForm.ptinv

NonlinearNormalForm.DAMap — Type

DAMap{S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix{S},Nothing}} <: TaylorMap{S,T,U,V}

TaylorMap that composes and inverses as a DAMap (with the scalar part ignored). See TaylorMap for more information.

NonlinearNormalForm.DAMap — Method

DAMap(M; x0::Vector{S}=zeros(numtype(first(x)), numvars(x)), Q::U=nothing, E::V=nothing,  spin::Union{Bool,Nothing}=nothing, radiation::Union{Bool,Nothing}=nothing, use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

M must represent a matrix with linear indexing.

Constructs a DAMap with the passed matrix of scalars M as the linear part of the TaylorMap, and optionally the entrance coordinates x0, Quaternion for spin Q, and stochastic matrix E as keyword arguments. The helper keyword arguments spin and radiation may be set to true to construct a DAMap with an identity quaternion/stochastic matrix, or false for no spin/radiation. Note that setting spin/radiation to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor.

NonlinearNormalForm.DAMap — Method

DAMap(p::Probe{S,T,U,V}; use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Creates a DAMap from the Probe, which must contain TPSs.

If use is not specified, then the same GTPSA Descriptor as p will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the p promoted to a DAMap will have the same Descriptor as in use. The number of variables and parameters must agree, however the orders may be different.

NonlinearNormalForm.DAMap — Method

DAMap(TaylorMap{S,T,U,V}; use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Creates a new copy of the passed TaylorMap as a DAMap.

If use is not specified, then the same GTPSA Descriptor as m will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the copy of m as a new DAMap will have the same Descriptor as in use. The number of variables and parameters must agree, however the orders may be different.

NonlinearNormalForm.DAMap — Method

DAMap(u::UndefInitializer, m::TaylorMap{S,T,U,V}) where {S,T,U,V}

Creates an undefined DAMap based on m (same Descriptor and with radiation/spin on/off.)

NonlinearNormalForm.DAMap — Method

DAMap(x::Vector{T}=zeros(TPS, numvars(GTPSA.desc_current)); x0::Vector{S}=zeros(numtype(first(x)), numvars(x)), Q::U=nothing, E::V=nothing,  spin::Union{Bool,Nothing}=nothing, radiation::Union{Bool,Nothing}=nothing, use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Constructs a DAMap with the passed vector of TPS/ComplexTPS as the orbital ray, and optionally the entrance coordinates x0, Quaternion for spin Q, and stochastic matrix E as keyword arguments. The helper keyword arguments spin and radiation may be set to true to construct a DAMap with an identity quaternion/stochastic matrix, or false for no spin/radiation. Note that setting spin/radiation to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor. The use kwarg may also be used to change the Descriptor of the TPSs, provided the number of variables and parameters agree (orders may be different).

NonlinearNormalForm.Quaternion — Type

Quaternion{T <: Number}

Lightweight quaternion implementation for simulations.

NonlinearNormalForm.S — Type

Generic symplectic skew symmetric S matrix (size inferred from other matrix in operations) using SkewLinearAlgebra's JMatrix

NonlinearNormalForm.SymplecticS — Type

Generic symplectic skew symmetric S matrix (size inferred from other matrix in operations) using SkewLinearAlgebra's JMatrix

NonlinearNormalForm.TPSAMap — Type

TPSAMap{S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix{S},Nothing}} <: TaylorMap{S,T,U,V}

TaylorMap that composes and inverses as a TPSAMap (with the scalar part included). See TaylorMap for more information.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(M; x0::Vector{S}=zeros(numtype(first(x)), numvars(x)), Q::U=nothing, E::V=nothing,  spin::Union{Bool,Nothing}=nothing, radiation::Union{Bool,Nothing}=nothing, use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

M must represent a matrix with linear indexing.

Constructs a TPSAMap with the passed matrix of scalars M as the linear part of the TaylorMap, and optionally the entrance coordinates x0, Quaternion for spin Q, and stochastic matrix E as keyword arguments. The helper keyword arguments spin and radiation may be set to true to construct a TPSAMap with an identity quaternion/stochastic matrix, or false for no spin/radiation. Note that setting spin/radiation to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(p::Probe{S,T,U,V}; use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Creates a TPSAMap from the Probe, which must contain TPSs.

If use is not specified, then the same GTPSA Descriptor as p will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the p promoted to a TPSAMap will have the same Descriptor as in use. The number of variables and parameters must agree, however the orders may be different.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(TaylorMap{S,T,U,V}; use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Creates a new copy of the passed TaylorMap as a TPSAMap.

If use is not specified, then the same GTPSA Descriptor as m will be used. If use is specified (could be another Descriptor, TaylorMap, or a Probe containing TPSs), then the copy of m as a new TPSAMap will have the same Descriptor as in use. The number of variables and parameters must agree, however the orders may be different.

NonlinearNormalForm.TPSAMap — Method

TPSAMap(u::UndefInitializer, m::TaylorMap{S,T,U,V}) where {S,T,U,V}

Creates an undefined TPSAMap based on m (same Descriptor and with radiation/spin on/off.)

NonlinearNormalForm.TPSAMap — Method

TPSAMap(x::Vector{T}=zeros(TPS, numvars(GTPSA.desc_current)); x0::Vector{S}=zeros(numtype(first(x)), numvars(x)), Q::U=nothing, E::V=nothing,  spin::Union{Bool,Nothing}=nothing, radiation::Union{Bool,Nothing}=nothing, use::Union{Descriptor,TaylorMap,Probe{S,T,U,V},Nothing}=nothing) where {S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix,Nothing}}

Constructs a TPSAMap with the passed vector of TPS/ComplexTPS as the orbital ray, and optionally the entrance coordinates x0, Quaternion for spin Q, and stochastic matrix E as keyword arguments. The helper keyword arguments spin and radiation may be set to true to construct a TPSAMap with an identity quaternion/stochastic matrix, or false for no spin/radiation. Note that setting spin/radiation to any Bool value without Q or E specified is type-unstable. This constructor also checks for consistency in the length of the orbital ray and GTPSA Descriptor. The use kwarg may also be used to change the Descriptor of the TPSs, provided the number of variables and parameters agree (orders may be different).

NonlinearNormalForm.TaylorMap — Type

TaylorMap{S,T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing},V<:Union{Matrix{S},Nothing}}

Abstract type for TPSAMap and DAMap used for normal form analysis.

All TaylorMaps contain x0 and x as the entrance coordinates and transfer map as a truncated power series respectively. If spin is included, a field Q containing a Quaternion as a truncated power series is included, else Q is nothing. If radiation is included, a field E contains a matrix of the envelope for stochastic radiation, else E is nothing.

Fields

x0 – Entrance coordinates of the map, Taylor expansion point

x – Orbital ray as a truncated power series, expansion around x0 + scalar part equal to EXIT coordinates of map

Q – Quaternion as a truncated power series if spin is included, else nothing

E – Matrix of the envelope for stochastic radiation if included, else nothing

NonlinearNormalForm.VectorField — Type

Lie operator to act on maps. Can be turned into a map with exp(:F:)

NonlinearNormalForm.VectorField — Method

VectorField(h::T; Q::U=nothing, spin::Union{Bool,Nothing}=nothing, work_low::Vector{<:Union{Ptr{RTPSA},Ptr{CTPSA}}}=Vector{lowtype(h)}(undef,numvars(h))) where {T<:Union{TPS,ComplexTPS},U<:Union{Quaternion{T},Nothing}}

Constructs a VectorField from the passed Hamiltonian h. Explicity, for h, constructs a vector field F such that

F.x = [∂h/∂p₁, -∂h/∂q₁, ...]

Base.:∘ — Method

∘(m2::DAMap,m1::DAMap) -> DAMap

DAMap composition, ignoring the scalar part of m1

Base.:∘ — Method

∘(m2::TPSAMap,m1::TPSAMap) -> TPSAMap

TPSAMap composition, including the scalar part of m1

Base.inv — Method

inv(m1::TaylorMap{S,T,U,V}; dospin::Bool=true, work_low::Tuple{Vararg{Vector{<:Union{Ptr{RTPSA},Ptr{CTPSA}}}}}=prep_inv_work_low(m1)) where {S,T,U,V}

Inverts the TaylorMap.

Keyword Arguments

dospin – Specify whether to invert the quaternion as well, default is true

work_low – Temporary vector to hold the low-level C pointers. Default is output from prep_inv_work_low

Base.log — Function

logpb(mf::Vector{<:Union{TPS,ComplexTPS}}, mi::Vector{<:Union{TPS,ComplexTPS}}=vars(first(F)))

Given a final map mf and initial map mi, this function calculates the vector field F such that mf=exp(F⋅∇)mi. If mi is not provided, it is assumed to be the identity.

julia> d = Descriptor(2,10); x = vars()[1]; p = vars()[2];

julia> time = 0.01; k = 2; m = 0.01;

julia> h = p^2/(2m) + 1/2*k*x^2;

julia> hf = getvectorfield(h);

julia> map = exppb(-time*hf);

julia> logpb(map) == -time*hf
true

Base.one — Method

one(m::DAMap{S,T,U,V}) where {S,T,U,V}

Construct an identity map based on m.

Base.one — Method

one(m::TPSAMap{S,T,U,V}) where {S,T,U,V}

Construct an identity map based on m.

Base.zero — Method

zero(m::DAMap{S,T,U,V}) where {S,T,U,V}

Creates a DAMap with the same GTPSA Descriptor, and spin/radiation on/off, as m but with all zeros for each quantity (except for the immutable parameters in x[nv+1:nn], which will be copied from m.x)

Base.zero — Method

zero(m::TPSAMap{S,T,U,V}) where {S,T,U,V}

Creates a TPSAMap with the same GTPSA Descriptor, and spin/radiation on/off, as m but with all zeros for each quantity (except for the immutable parameters in x[nv+1:nn], which will be copied from m.x)

GTPSA.compose! — Method

compose!(m::DAMap, m2::DAMap, m1::DAMap; dospin::Bool=true, keep_scalar::Bool=true, work_ref::Union{Nothing,Vector{<:Union{Float64,ComplexF64}}}=nothing, work_low::Tuple{Vararg{Vector{<:Union{Ptr{RTPSA},Ptr{CTPSA}}}}}=prep_comp_work_low(m), work_prom::Union{Nothing,Tuple{Vararg{Vector{<:ComplexTPS}}}}=prep_comp_work_prom(m,m2,m1))

In-place DAMap composition which calculates m = m2 ∘ m1, ignoring the scalar part of m1.

Aliasing m === m2 is allowed, but NOT m === m1. m2 === m1 is fine. The destination map m should be properly set up (with correct types promoted if necessary), and have m.x[1:nv] (and m.Q.q if spin) containing allocated TPSs.

If keep_scalar is set to true, then the scalar part of m1 is retained. In this case, a temporary vector work_ref must be used to store the scalar part before calling the low-level compose_it!, then added back into m1 after composition. work_ref can be optionally provided, or will be created internally in this case. Default is true.

If dospin is true, then the quaternion part of the maps will be composed as well. Default is true.

See the documentation for compose_it! for information on work_low and work_prom.

Keyword Arguments

keep_scalar – Specify whether to keep the scalar part in m1 or throw it away. If true, a temporary vector storing the scalar part must be used. Default is true

work_ref – If keep_scalar is true, the temporary vector can be provided in this keyword argument. If nothing is provided, the temporary will be created internally. Default is nothing

dospin – Specify whether or not to include the quaternions in the concatenation. Default is true

work_low – Temporary vector to hold the low-level C pointers. See the compose_it! documentation for more details. Default is output from prep_comp_work_low(m)

work_prom – Temporary vector of allocated ComplexTPSs when there is implicit promotion. See the compose_it! documentation for more details. Default is output from prep_comp_work_prom(m, m2, m1)

GTPSA.compose! — Method

compose!(m::TPSAMap, m2::TPSAMap, m1::TPSAMap; dospin::Bool=true, work_low::Tuple{Vararg{Vector{<:Union{Ptr{RTPSA},Ptr{CTPSA}}}}}=prep_comp_work_low(m), work_prom::Union{Nothing,Tuple{Vararg{Vector{<:ComplexTPS}}}}=prep_comp_work_prom(m,m2,m1))

In-place TPSAMap composition which calculates m = m2 ∘ m1, including the scalar part of m1.

Aliasing m === m2 is allowed, but NOT m === m1. m2 === m1 is fine. The destination map m should be properly set up (with correct types promoted if necessary), and have m.x[1:nv] (and m.Q.q if spin) containing allocated TPSs.

If dospin is true, then the quaternion part of the maps will be composed as well. Default is true.

See the documentation for compose_it! for information on work_low and work_prom.

Keyword Arguments

dospin – Specify whether or not to include the quaternions in the concatenation. Default is true

work_low – Temporary vector to hold the low-level C pointers. See the compose_it! documentation for more details. Default is output from prep_comp_work_low(m)

work_prom – Temporary vector of allocated ComplexTPSs when there is implicit promotion. See the compose_it! documentation for more details. Default is output from prep_comp_work_prom(m, m2, m1)

LinearAlgebra.mul! — Method

mul!(Q3::Quaternion, Q1::Quaternion{S}, Q2::Quaternion{T}) where {S,T}

Multiplies the quaternions Q1*Q2 and stores the result in-place in Q3. Aliasing of all three arguments is allowed.

LinearAlgebra.mul! — Method

NonlinearNormalForm.checksymp — Method

checksymp(M)

Returns tranpose(M)*S*M - S, where S is the skew-symmetric matrix S = blkdiag([0 1; -1 0], ...). If M is symplectic, then the result should be a matrix containing all zeros. The non-symplectic parts of the matrix can be identified by those nonzero elements in the result.

NonlinearNormalForm.compose — Method

compose(m2::DAMap,m1::DAMap) -> DAMap

DAMap composition, which calculates m2 ∘ m1 ignoring the scalar part of m1

NonlinearNormalForm.compose — Method

compose(m2::TPSAMap,m1::TPSAMap) -> TPSAMap

TPSAMap composition, which calculates m2 ∘ m1 including the scalar part of m1

NonlinearNormalForm.compose_it! — Method

compose_it!(m, m2, m1; dospin=true, work_low=nothing, work_prom=nothing)

Low level composition function, m = m2 ∘ m1. Aliasing m with m2 is allowed, but not m with m1. Assumes the destination map is properly set up (with correct types promoted if necessary), and that m.x[1:nv] (and m.Q.q if spin) contain allocated TPSs. The parameters part of m.x (m.x[nv+1:nn]) does not need to contain allocated TPSs.

For all compositions, 5 temporary vectors must be generated that contain Ptr{RTPSA} or Ptr{CTPSA} for each TPS in the map (depending on output type), to pass to the low-level C composition function in GTPSA. Three vectors are for the orbital part (m.x, m2.x, m1.x correspondingly referred in the code as outx_low, m2x_low, and m1x_low) and two are for the spin part (m.Q.q and m2.Q.q correspondingly referred in the code as outQ_low and m2Q_low). These 5 temporaries can be optionally passed as a tuple in work_low, and must satisfy the following requirements:

work_low[1] = outx_low   # Length >= nv
work_low[2] = m2x_low    # Length >= nv
work_low[3] = m1x_low    # Length >= nv+np
work_low[4] = outQ_low   # Length >= 4, could be = work_low[1] or work_low[2] if nv >= 4
work_low[5] = m2Q_low    # Length >= 4, could be = work_low[1] or work_low[2] if nv >= 4

Furthermore, for the spin part both of outx_low and m2x_low could be reused for outQ_low and m2Q_low if nv >= 4 , however m1x_low MUST NOT BE REUSED!

If promotion is occuring, then one of the maps is ComplexTPS and the other TPS, with output map ComplexTPS. Note that the spin part is required to agree with the orbital part in terms of type by definition of the TaylorMap struct. work_prom can optionally be passed as a tuple containing the temporary ComplexTPSs if promotion is occuring:

If eltype(m.x) != eltype(m1.x) (then m1 must be promoted): work_prom[1] = m1x_prom # Length >= nv+np, Vector{ComplexTPS}

else if eltype(m.x) != eltype(m2.x) (then m2 must be promoted):

work_prom[1] = m2x_prom  # Length >= nv, Vector{ComplexTPS}
work_prom[2] = m2Q_prom  # Length >= 4, Vector{ComplexTPS}

Note that the ComplexTPSs in the vector(s) must be allocated and have the same Descriptor.

If spin is included, not that the final quaternion concatenation step mul! will creat allocations

NonlinearNormalForm.compose_it! — Method

compose_it!(m, m2, m1; dospin=true, work_low=nothing, work_prom=nothing)

Low level composition function, m = m2 ∘ m1. Aliasing m with m2 is allowed, but not m with m1. Assumes the destination map is properly set up (with correct types promoted if necessary), and that m.x[1:nv] (and m.Q.q if spin) contain allocated TPSs. The parameters part of m.x (m.x[nv+1:nn]) does not need to contain allocated TPSs.

For all compositions, 5 temporary vectors must be generated that contain Ptr{RTPSA} or Ptr{CTPSA} for each TPS in the map (depending on output type), to pass to the low-level C composition function in GTPSA. Three vectors are for the orbital part (m.x, m2.x, m1.x correspondingly referred in the code as outx_low, m2x_low, and m1x_low) and two are for the spin part (m.Q.q and m2.Q.q correspondingly referred in the code as outQ_low and m2Q_low). These 5 temporaries can be optionally passed as a tuple in work_low, and must satisfy the following requirements:

work_low[1] = outx_low   # Length >= nv
work_low[2] = m2x_low    # Length >= nv
work_low[3] = m1x_low    # Length >= nv+np
work_low[4] = outQ_low   # Length >= 4, could be = work_low[1] or work_low[2] if nv >= 4
work_low[5] = m2Q_low    # Length >= 4, could be = work_low[1] or work_low[2] if nv >= 4

Furthermore, for the spin part both of outx_low and m2x_low could be reused for outQ_low and m2Q_low if nv >= 4 , however m1x_low MUST NOT BE REUSED!

If promotion is occuring, then one of the maps is ComplexTPS and the other TPS, with output map ComplexTPS. Note that the spin part is required to agree with the orbital part in terms of type by definition of the TaylorMap struct. work_prom can optionally be passed as a tuple containing the temporary ComplexTPSs if promotion is occuring:

If eltype(m.x) != eltype(m1.x) (then m1 must be promoted): work_prom[1] = m1x_prom # Length >= nv+np, Vector{ComplexTPS}

else if eltype(m.x) != eltype(m2.x) (then m2 must be promoted):

work_prom[1] = m2x_prom  # Length >= nv, Vector{ComplexTPS}
work_prom[2] = m2Q_prom  # Length >= 4, Vector{ComplexTPS}

Note that the ComplexTPSs in the vector(s) must be allocated and have the same Descriptor.

If spin is included, not that the final quaternion concatenation step mul! will creat allocations

NonlinearNormalForm.exppb — Function

exppb(F::Vector{<:Union{TPS,ComplexTPS}}, m::Vector{<:Union{TPS,ComplexTPS}}=vars(first(F)))

Calculates exp(F⋅∇)m = m + F⋅∇m + (F⋅∇)²m/2! + .... If m is not provided, it is assumed to be the identity.

Example

julia> d = Descriptor(2,10); x = vars()[1]; p = vars()[2];

julia> time = 0.01; k = 2; m = 0.01;

julia> h = p^2/(2m) + 1/2*k*x^2;

julia> hf = getvectorfield(h);

julia> map = exppb(-time*hf, [x, p])
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:   9.9001665555952290e-01      1      1   0
   1:   9.9666999841313930e-01      1      0   1
-------------------------------------------------
   2:  -1.9933399968262787e-02      1      1   0
   2:   9.9001665555952378e-01      1      0   1

NonlinearNormalForm.fgrad — Method

fgrad(F::Vector{<:T}, h::T) where {T<:Union{TPS,ComplexTPS}}

Calculates F⋅∇h.

NonlinearNormalForm.gethamiltonian — Method

gethamiltonian(F::Vector{<:Union{TPS,ComplexTPS}})

Assuming the variables in the TPSA are canonically-conjugate, and ordered so that the canonically- conjugate variables are consecutive (q₁, p₁, q₂, p₂, ...), this function calculates the Hamiltonian from a vector field F that can be obtained from a Hamiltonian (e.g. by getvectorfield). Explicitly, ∫ F₁ dp₁ - ∫ F₂ dq₁ + ... + ∫ F₂ₙ₋₁ dpₙ - ∫ F₂ₙ dqₙ

Example

julia> d = Descriptor(2,10); x = vars();

julia> h = (x[1]^2 + x[2]^2)/2
TPS:
 Coefficient                Order   Exponent
  5.0000000000000000e-01      2      2   0
  5.0000000000000000e-01      2      0   2

julia> F = getvectorfield(h)
2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-------------------------------------------------
   1:  -1.0000000000000000e+00      1      0   1
-------------------------------------------------
   2:   1.0000000000000000e+00      1      1   0

julia> gethamiltonian(F)
TPS:
 Coefficient                Order   Exponent
  5.0000000000000000e-01      2      2   0
  5.0000000000000000e-01      2      0   2

NonlinearNormalForm.inv! — Method

inv!(m::TaylorMap{S,T,U,V}, m1::TaylorMap{S,T,U,V}; dospin::Bool=true, work_ref::Union{Nothing,Vector{<:Union{Float64,ComplexF64}}}=nothing, work_low::Tuple{Vararg{Vector{<:Union{Ptr{RTPSA},Ptr{CTPSA}}}}}=prep_inv_work_low(m1)) where {S,T,U,V}

In-place inversion of the TaylorMap setting m = inv(m1). Aliasing m === m1 is allowed, however in this case a temporary vector must be used to store the scalar part of m1 prior to inversion so that the entrance/exit coordinates of the map can be properly handled.

Keyword Arguments

dospin – Specify whether to invert the quaternion as well, default is true

work_ref – If m === m1, then a temporary vector must be used to store the scalar part. If not provided and m === m1, this temporary will be created internally. Default is nothing

work_low – Temporary vector to hold the low-level C pointers. Default is output from prep_inv_work_low

NonlinearNormalForm.lb! — Method

lb!(G::VectorField{T,U}, F::VectorField{T,U}, H::VectorField{T,U}; work_low::Tuple{Vararg{Vector{<:Union{Ptr{RTPSA},Ptr{CTPSA}}}}}=prep_vf_work_low(F), work_Q::Union{Nothing,U}=nothing) where {T,U}

Computes the Lie bracket of the vector fields F and H, and stores the result in G. See lb for more details.

Keyword Arguments

work_low – Tuple of 3 vectors of type lowtype(T) of length >=nv

work_Q – Quaternion{T} if spin is included in the vector field, else nothing

NonlinearNormalForm.lb — Method

lb(F::VectorField{T,U}, H::VectorField{T,U}) where {T,U}

Computes the Lie bracket of the vector fields F and H. Explicitly, and including spin (with the lower case letter for the quaternion of the vector field), this is:

(G,g) = ⟨(F,f) , (H,h)⟩ = (F⋅∇H-H⋅∇F , [h,f]+F⋅∇h-G⋅∇f)

where [h,f] = h*f - f*h is just a quaternion commutator. See Equation 44.52 in the Bmad manual

NonlinearNormalForm.locate_modes! — Function

locate_modes!(evecs, evals=nothing; sort=true, modes=nothing)

For each mode (every pair of rows in evecs), determines the eigenvector (each column in evecs) with the largest norm in that mode. The eigenvectors must have dimensionality of 2n, n ∈ ℤ and be next to each other pairs.

If sort is true, then the pairs of eigenvectors are sorted in-place. The eigenvalues will also be sorted if evals is provided. If sort is false, then the modes vector is filled with a mapping of the current eigenvector position to its mode (index of modes is the current eigenvector position, and the element at that index in modes is the mode that it corresponds to). For sort=true, modes == 1:n after calling this function.

If the mode locating is successful, true is returned. For unsuccessful mode locating, false is returned, modes will contain junk, and the eigenvectors/values may be partially sorted.

Assumes the eigenvectors (and eigenvalues if provided) are already in pairs (next to each other) starting at the first column.

Input:

evecs – 2n x 2n matrix of eigenvectors already in pairs

evals – Vector of length 2n containing the eigenvalues corresponding to the eigenvectors in evecs

sort – (Optional) kwarg to specify whether or not to sort the eigenvectors, default is true

Output:

ret – Returns true if the mode locating is successful, false if otherwise

modes – (Optional) kwarg for eigvec -> mode mapping, length(evecs) == Int(size(evecs,1)/2)

NonlinearNormalForm.log! — Method

Instead of using GTPSA's logpb, we need a log that will work with VectorFields including a quaternion. Therefore I am writing my own.

If we formulate the matrix of monomials, the log is easy. But this is gross and slow.

Therefore we use the fast algorithm described in Ch. 47 of Bmad manual

NonlinearNormalForm.log! — Method

All arguments must have correct types. No promotion here yet. I am tired. Maybe soon

Keyword arguments

'work_low – Tuple of 3 vectors of length nv with type equal to the lowtype of F.

NonlinearNormalForm.mat_eigen! — Method

mat_eigen!(mat; sort=true, phase_modes=true)

Same as mat_eigen, but mutates mat for speed. See the documentation for mat_eigen for more details.

NonlinearNormalForm.mat_eigen — Method

mat_eigen(mat; sort=true, phase_modes=true)

Given a matrix mat with an even number of rows/cols, calculates the eigenvectors and eigenvalues. For stable eigenmodes (complex conjugate eigenvector/eigenvalue pairs), the eigenvectors are normalized so that vⱼ'*S*vⱼ = +im for odd j, and -im for even j.

If sort is true, then each eigenvector/eigenvalue pair will be sorted according to the mode it best identifies with. A warning will be printed if the mode sorting fails. Mode sorting will automatically fail if more than 1 mode is unstable. Default is true.

If phase_modes is true, then each stable mode will be multiplied by a phase factor to make the first component of that mode (e.g. the x, y, or z component, NOT the px, py, or pz component) be fully real. This both makes the eigenvector "pretty", and in this case of weak coupling, ensures that vⱼ'*vⱼ₊₁ for j=(1, 2, 3) has a positive imaginary part for odd j so that the eigenvectors/ eigenvalues for odd j are associated with the positive tune and even j the negative tune. This phase factor is harmless/useless to include in a highly-coupled matrix.

For complex matrices, Julia's eigen, which is called by mat_eigen, is type-unstable.

NonlinearNormalForm.moveback_unstable! — Method

moveback_unstable!(F::Eigen) -> Int

Moves back eigenvectors with eigenvalues having a zero imaginary component to the end of the values and vectors arrays in the Eigen struct, and returns the number of unstable eigenvectors.

Note that if more than 1 mode is unstable, the pair of eigenvectors corresponding to a mode will not necessarily be next to each other at the end of the matrix.

NonlinearNormalForm.normalize_eigenmode! — Function

normalize_eigenmode!(evec_pair, eval_pair, phase_mode::Integer=-1)

Normalizes the complex-conjugate eigenvector pair so that vⱼ'*S*vⱼ = +im for j=1, and -im for j=2. This may involve swapping the complex-conjugate eigenvectors/eigenvalues.

If phase_mode is set to a particular mode (1, 2, 3), then a phase will be multiplied by the eigenvectors to make them "pretty"; the phase will make the first component of that mode (e.g. the x, y, or z component) be fully real. This gives the eigenvectors a simple form in the weakly-coupled case, with the product v₁'*v₂ having a positive imaginary part, and ensures the odd numbered eigenvector/eigenvalues are associated with the tune and the even numbered ones have the negative of the tune. In cases with a lot of coupling, there is no simple way to define the positive or negative tune, and multiplying by this phase factor is harmless. See the "Tunes From One-Turn Matrix Eigen Analysis" section in the Bmad manual for more details.

NonlinearNormalForm.pb! — Method

pb!(h::Union{TPS,ComplexTPS}, f::Union{TPS,ComplexTPS}, g::Union{TPS,ComplexTPS})

In-place Poisson bracket, see the documentation for pb for more details.

NonlinearNormalForm.pb — Method

pb(f::Union{TPS, ComplexTPS}, g::Union{TPS, ComplexTPS})

Assuming the variables in the TPSA are canonically-conjugate, and ordered so that the canonically- conjugate variables are consecutive (q₁, p₁, q₂, p₂, ...), computes the Poisson bracket of the scalar functions f and g. The Poisson bracket of two functions {f, g} is defined as Σᵢ (∂f/∂qᵢ)(∂g/∂pᵢ) - (∂g/∂qᵢ)(∂f/∂pᵢ).

Examples

julia> d = Descriptor(4,10);

julia> x = vars(d);

julia> f = (x[1]^2 + x[2]^2)/2 + (x[3]^2 + x[4]^2)/2;

julia> pb(f,x[1])
TPS:
  Coefficient              Order     Exponent
  -1.0000000000000000e+00    1        0    1    0    0

julia> pb(f,x[2])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        1    0    0    0

julia> pb(f,x[3])
TPS:
  Coefficient              Order     Exponent
  -1.0000000000000000e+00    1        0    0    0    1

julia> pb(f,x[4])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        0    0    1    0

NonlinearNormalForm.ptinv — Method

ptinv(ma::Vector{<:Union{TPS,ComplexTPS}}, vars::Vector{<:Integer})

Partially-inverts the map ma, inverting only the variables specified by index in vars.